requirements.txt
alembic==1.3.0
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.2.0
Automat==0.6.0
awscli==1.16.272
Babel==2.6.0
bcrypt==3.1.6
beautifulsoup4==4.8.1
blinker==1.4
blis==0.4.1
boto3==1.10.8
botocore==1.13.8
bs4==0.0.1
bsddb3==6.2.6
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
CherryPy==8.9.1
Click==7.0
cloud-init==19.2
colorama==0.4.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.6.1
cymem==2.0.2
dbus-python==1.2.12
decorator==4.3.0
distro==1.3.0
distro-info===0.21ubuntu4
dj-database-url==0.5.0
Django==2.2.6
django-extensions==2.2.5
django-heroku==0.3.1
djangorestframework==3.10.3
docutils==0.15.2
en-core-web-sm==2.2.0
entrypoints==0.3
et-xmlfile==1.0.1
fanfiction==0.1.1
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.18.2
gevent==1.4.0
goodreads==0.3.2
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.21.0
google-cloud-texttospeech==0.5.0
google-compute-engine==20190801.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
greenlet==0.4.15
grequests==0.4.0
grpcio==1.24.3
gunicorn==19.9.0
Gutenberg==0.8.0
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.11.3
hyperlink==17.3.1
idna==2.8
importlib-metadata==0.23
incremental==16.10.1
ipython==5.8.0
ipython-genutils==0.2.0
ipython3==0.0.1
isodate==0.6.0
itsdangerous==0.24
jdcal==1.0
Jinja2==2.10
jmespath==0.9.4
jsonpatch==1.22
jsonpickle==1.2
jsonpointer==1.10
jsonschema==2.6.0
keyring==18.0.1
keyrings.alt==3.1.1
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
murmurhash==1.0.2
netifaces==0.10.4
nose==1.3.7
numpy==1.17.3
oauthlib==2.1.0
olefile==0.46
openpyxl==2.4.9
pandas==0.25.3
paramiko==2.6.0
passlib==1.7.1
pexpect==4.6.0
pgspecial==1.9.0
pickleshare==0.7.5
Pillow==6.1.0
plac==1.1.3
pluggy==0.12.0
preshed==3.0.2
prettytable==0.7.2
prompt-toolkit==1.0.15
protobuf==3.10.0
psutil==5.5.1
psycopg2==2.8.4
py==1.8.0
pyaes==1.6.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycrypto==2.6.1
Pygments==2.3.1
PyGObject==3.34.0
pyinotify==0.9.6
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyserial==3.4
pytest==3.10.1
python-apt==1.9.0+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.36
python-editor==1.0.4
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
rauth==0.7.3
rdflib==4.2.2
rdflib-sqlalchemy==0.3.8
repoze.lru==0.7
requests==2.22.0
requests-unixsocket==0.1.5
RestrictedPython==4.0b3
roman==2.0.0
Routes==2.4.1
rsa==3.4.2
s3transfer==0.2.1
scapy==2.4.3
scrypt==0.8.0
SecretStorage==2.3.1
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.5
spacy==2.2.2
SPARQLWrapper==1.8.4
SQLAlchemy==1.3.10
sqlparse==0.3.0
srsly==0.2.0
ssh-import-id==5.7
sshtunnel==0.1.4
systemd-python==234
thinc==7.3.1
tornado==5.1.1
tqdm==4.37.0
traitlets==4.3.2
Twisted==18.9.0
ubuntu-advantage-tools==19.5
ufw==0.36
unattended-upgrades==0.1
untangle==1.1.1
urllib3==1.25.6
wadllib==1.3.3
wasabi==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==0.14.1
whitenoise==4.1.4
WTForms==2.2.1
xmltodict==0.12.0
zipp==0.6.0
zope.interface==4.3.2
